refactor(cli): split build_profile into six concern modules
Decompose the 1,235-line build_profile.py into sibling modules —
schema (config dataclasses), presets (bundled preset/trigger
discovery), merge (extends/deep-merge/exclude + resolved-profile
hashing), model (BuildProfile + validate), load (YAML -> BuildProfile),
resolve (CLI-facing resolution) — behind a pure re-export facade with a
sorted __all__ covering the grep-derived consumer surface.

Behavior is unchanged: signatures, error wording, and preset hash
digests are identical; module loggers now use __name__, with the two
pinned caplog sites pointed at the emitting modules. Namespace-coupled
test sites (monkeypatches, source-text guards, stale comments) migrate
to the module that owns each seam, per-concern edge-case tests land in
the owning suites, and three doc pointers now cite the new module
homes.
