+++ tests/cli/test_build_extends_preset.py
@@ -12,8 +12,10 @@ from pathlib import Path
 import pytest
 from click.testing import CliRunner
 
+from osprey.cli import build_profile_presets
 from osprey.cli.build_cmd import build
 from osprey.cli.build_profile import (
+    _load_preset_raw,
     _preset_exists,
     _presets_dir,
     list_presets,
@@ -164,3 +166,33 @@ def test_presets_dir_returns_directory_with_known_presets() -> None:
     assert presets.is_dir()
     files = {p.name for p in presets.glob("*.yml")}
     assert "hello-world.yml" in files
+
+
+def test_load_preset_raw_invalid_yaml_raises_build_profile_error(
+    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
+) -> None:
+    """A preset file with invalid YAML raises `BuildProfileError` naming the preset."""
+    from osprey.errors import BuildProfileError
+
+    monkeypatch.setattr(build_profile_presets, "_presets_dir", lambda: tmp_path)
+    (tmp_path / "broken.yml").write_text("name: [unterminated\n", encoding="utf-8")
+    with pytest.raises(BuildProfileError) as exc:
+        _load_preset_raw("broken")
+    msg = str(exc.value)
+    assert "Invalid YAML" in msg
+    assert "broken" in msg
+
+
+def test_load_preset_raw_non_mapping_yaml_raises_build_profile_error(
+    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
+) -> None:
+    """A preset file whose YAML parses to a list (not a mapping) is rejected."""
+    from osprey.errors import BuildProfileError
+
+    monkeypatch.setattr(build_profile_presets, "_presets_dir", lambda: tmp_path)
+    (tmp_path / "listy.yml").write_text("- one\n- two\n", encoding="utf-8")
+    with pytest.raises(BuildProfileError) as exc:
+        _load_preset_raw("listy")
+    msg = str(exc.value)
+    assert "must be a YAML mapping" in msg
+    assert "listy" in msg
